+++ services/horizon/CHANGELOG.md
@@ -11,6 +11,7 @@ bumps.  A breaking change will get clearly notified in this log.
 ### Added
 
 - Operation and payment resources were changed to add a `transaction_hash` property.
+- The ledger resource was changed to add a `header_xdr` property.  Existing horizon installations should re-ingest all ledgers to populate the history database tables with the data.  In future versions of horizon we will disallow null values in this column.  Going forward, this change reduces the coupling of horizon to stellar-core, ensuring that horizon can re-import history even when the data is no longer stored within stellar-core's database.
 
 ### Changed